fix(movement): widen tile friction to u16 and match the server's rule
Ten ground items in the shared asset data carry friction above 255. The
client truncated every friction through a `u8` at parse time, so on those
tiles it predicted a step 3-10x too short, fired its next walk early, and
was refused by the server every time.

Widen the field, the map lookup and `get_step_duration` to `u16`, and
restrict the parse to ground items so `get_tile_friction` can adopt the
server's "first item carrying a friction value" rule instead of filtering
on the Ground flag. The two selections are equivalent on the real data;
expressing the rule the same way on both sides keeps them from drifting.

The minimap stays `u8` and clamps at 254: chunks persist one byte per
tile with 0xFF reserved for "unvisited", and widening the format would
invalidate every player's explored map. It only uses friction for A*
passability and relative cost.

Adds `step_duration_matches_the_server`, the counterpart to the server's
`walk_ticks_match_the_client`. The formula is a deliberate duplicate
across two repositories, so each side pins the same three answers.

+++ src/items/item.rs
@@ -122,7 +122,7 @@ pub enum ItemFlag {
 pub struct ItemConfig {
     pub id: ItemId,
     pub flags: Vec<ItemFlag>,
-    pub friction: Option<u8>,
+    pub friction: Option<u16>,
     pub slot: Option<InventorySlot>,
     pub minimap_color: Option<u8>,
     pub elevation: Option<u8>,
